refactor(ao): extract agents lint adapter
